Contain PoC query in standalone service
